add addFriend and deleteFriend resolvers
