+++ src/models/User.model.ts
@@ -5,6 +5,7 @@ const UserSchema = new mongoose.Schema({
   password: String,
   phoneNumber: String,
   email: String,
+  friends: [String],
 });
 
 const User = mongoose.model("User", UserSchema);
